+++ src/features/climate-feed/components/ClimateFeedCard.tsx
@@ -3,8 +3,7 @@ import { Image, Pressable, StyleSheet, View } from 'react-native';
 import { capitalizeFirstLetter } from 'src/utils';
 import ClimateEffect from 'src/types/ClimateEffect';
 import ActionCardHeader from './ActionCardHeader';
-import Card from 'src/components/Cards/Card';
-import { CmTypography, CmChip  } from '@shared/components';
+import { CmTypography, CmChip, Card  } from '@shared/components';
 
 interface Props {
   climateEffect: ClimateEffect;
