+++ src/features/conversations/components/ConversationsIntroCard.tsx
@@ -1,8 +1,7 @@
 import { useState } from 'react';
 import { Pressable, StyleSheet } from 'react-native';
 
-import { CmTypography } from '@shared/components';
-import Card from 'src/components/Cards/Card';
+import { CmTypography, Card } from '@shared/components';
 
 function ConversationsIntroCard() {
   const [expanded, setExpanded] = useState(true);
