+++ src/features/conversations/components/DeleteConversationModal.tsx
@@ -1,7 +1,6 @@
 import { Modal, Pressable, StyleSheet, View } from 'react-native';
 
-import { CmTypography } from '@shared/components';
-import Card from 'src/components/Cards/Card';
+import { CmTypography, Card } from '@shared/components';
 
 interface Props {
   show: boolean;
