+++ src/features/conversations/components/ViewSelectedTopicsModal.tsx
@@ -12,7 +12,7 @@ import SolutionsFeedCard from './SolutionCard';
 import SolutionDetailsModal from './SolutionDetailsModal';
 import ActionDetailsModal from './ActionDetailsModal';
 import { CmTypography } from '@shared/components';;
-import Content from 'src/components/Screen/Content';
+import { Content } from '@shared/components';
 
 interface Props {
   open: boolean;
