+++ src/features/profile/components/ChangePasswordModal.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { Modal, Pressable, StyleSheet, TextInput, View } from 'react-native';
 import useApiClient from 'src/hooks/useApiClient';
 import { CmTypography, showErrorToast, showSuccessToast } from '@shared/components';
-import Card from 'src/components/Cards/Card';
+import Card from 'src/shared/components/Cards/Card';
 
 interface Props {
   show: boolean;
