+++ src/features/profile/components/UpdateEmailModal.tsx
@@ -4,7 +4,7 @@ import { Modal, Pressable, StyleSheet, TextInput, View } from 'react-native';
 import { useAppSelector } from 'src/store/hooks';
 import useApiClient from 'src/hooks/useApiClient';
 import { CmTypography } from '@shared/components';
-import Card from 'src/components/Cards/Card';
+import Card from 'src/shared/components/Cards/Card';
 
 interface Props {
   show: boolean;
