+++ src/screens/UserAAuthorizedScreens/ClimateDetailsScreen/ClimateDetailsScreen.tsx
@@ -7,11 +7,8 @@ import { ClimateFeedStackParams } from 'src/navigation/Stacks/ClimateFeedStack';
 import { ActionCard } from 'src/features/climate-feed/components';
 
 import { capitalizeFirstLetter, openUrl } from 'src/utils';
-import Screen from 'src/components/Screen/Screen';
-import Section from 'src/components/Screen/Section';
-import Content from 'src/components/Screen/Content';
 
-import { CmTypography, DetailsSourcesTab, BackButton  } from '@shared/components';
+import { CmTypography, DetailsSourcesTab, BackButton, Screen, Section, Content } from '@shared/components';
 
 type Props = NativeStackScreenProps<ClimateFeedStackParams, 'ClimateDetailsScreen'>;
