+++ src/shared/components/Cards/PersonalValueCardSmall.tsx
